apps/api/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
redis==4.6.0
websockets==11.0.3
pydantic==2.3.0
langchain==0.0.275
langchain-openai==0.0.2
python-dotenv==1.0.0
tiktoken==0.4.0
